Add new emacs-gui-server and emacs-gui-client commands
The idea behind these commands is to easily launch a Emacs server which
runs in the background, but with font/theme/etc settings as if it was
launched like a normal GUI.

I used to just launch Emacs.app directly, but, now I use these binaries
to open a full GUI frame by executing `emacs-gui-client -c &`. After
starting the GUI server with `emacs-gui-server start` of course.

The main difference between calling `emacs --daemon` and
`emacs-gui-server start` is that the latter uses a custom socket as to
not interfere with any console-base Emacs server, and it also sets the
`EMACS_GUI_SERVER` environment variable to `1`, letting my Emacs config
figure out which theme/fonts/etc to load.