fix: ensure goenv is always loaded so GOPATH is correctly set
The lazy-loading could lead to incorrect GOPATH values in sub-shells and
tmux instances which inherited a old GOPATH value from the parent
shell.

This is mostly an issue when new Go versions are installed, but it's
still an issue. Hence I now always load goenv in the background ensuring
it sets GOPATH correctly regardless of any existing GOPATH value.