chore(zshenv): remove needless PATH setup
Path setup for shim dirs of rbenv, pyenv, and nodenv are no longer
needed, as their init scripts are cached and loaded in the
background. And their init scripts will add the shim directories to
PATH.

The setup we're removing here are left-over from when all three tools
were being lazy loaded upon first use of the command, or first shell
completion for the command.