fix(xbar/brew-updates): refresh xbar plugin after update in terminal completes
Previously the xbar plugin would refresh as soon as you clicked on an
update, but since brew upgrades takes much longer to run in the popped
up terminal window. This now uses the xbar:// URI scheme to trigger a
plugin refresh after the brew commands have run in the terminal.