fix(node/prettier): dirty hack to allow prettier to find plugins installed with volta
Volta installs all global packages into their own separate sandbox, with
no other packages installed in the same node_modules directory.

This hack abuses the fact that prettier can be configured with a *.js
file, and effectively scans the volta installation directory for
sandboxes for prettier plugins, and sets anything it finds as the
"pluginSearchDirs" list.