feat(hammerspoon/app_toggle): enable multi-app toggles
A multi-app toggle is a keybinding which is configured to toggle 2 or
more applications.

This is intended as a context-ish-aware toggle, as it will only toggle
the most recently focused application. This essentially enables you to
bind a category/class of applications to a single hotkey, and whichever
of the apps that's running and was most recently focused is the one that
will be toggled.