fix(vscode): avoid interactive shell setup during VSCode shell resolving
When VSCode is started from normal non-terminal/CLI methods, it will start a interactive shell session in the background to dump the the env and get PATH.

Latest versions of VSCode do no play nice with Mise's dynamic PATH updating stuff, and understandably so.

Hence we now check VSCODE_RESOLVING_ENVIRONMENT and avoid all interactive shell setup when it is set.